chore: move notify logic to notification service
