+++ src/utils/geolocation/buildLatLng.js
@@ -1,0 +1,6 @@
+const buildLatLong = (coords) => ({
+  longitude: coords[0],
+  latitude: coords[1],
+});
+
+module.exports = buildLatLong;
